Resolve the size-ratchet regression: extract pricing catalogs, shrink config
Review fix batch 2 (triad-codex critical, Phase A wave): the DeepSeek
provider work pushed two deterministic size gates over their ceilings.

- ouroboros/llm.py re-entered BYTE_DEBT (>200,000 UTF-8 bytes): both live
  pricing-catalog fetchers (OpenRouter + cloud.ru) move whole to the new
  leaf module ouroboros/provider_catalogs.py; llm.py re-exports the
  historical names so ouroboros.pricing and tests keep one import surface.
  llm.py is now 198,329 bytes.
- ouroboros/config.py crossed the 1600-line GIANT ceiling (1602): inline
  the deepseek presence probe into the exclusive-direct-provider tuple and
  tighten the routing comment. Back to exactly 1600 lines.
- Regenerate size_ratchet_manifest.py; tests/test_provider_contract_ci.py
  and web/modules/settings_ui.js entered the 1001-1500 band with the
  DeepSeek rows and carry explicit split-on-next-touch rationales.
- docs/ARCHITECTURE.md module map gains the provider_catalogs.py row in
  the same commit (checklist item 7).

pytest -m size_ratchet: 5 passed against both candidate PR bases
(origin/main and origin/ouroboros merge-bases via OURO_SIZE_RATCHET_BASE_REF).
