+++ docs/ARCHITECTURE.md
@@ -103,6 +103,7 @@ server.py (Starlette+uvicorn) ← HTTP + WebSocket on configurable host:port (de
       ├── code_intelligence.py ← Internal code inventory v2: derived-only file facts, hashes, polyglot symbol/import/call/reference extraction via tree-sitter for non-Python languages (Go/Rust/Java/Ruby/C/...) with Python on the stdlib `ast` path and a visible `structural_unavailable` fallback when a grammar is missing, plus an incremental JSON cache (no raw source)
       ├── code_search_rg.py    ← Optional ripgrep-backed search helper for search_code; every match is post-filtered through Ouroboros protected/secret gates
       ├── pricing.py           ← Exact-route best-effort provider-catalog lookup with nullable estimates; no static model tariffs and not the monetary ledger
+      ├── provider_catalogs.py ← Live pricing-catalog fetchers (OpenRouter + cloud.ru) extracted whole from `llm.py` at the 200,000-byte module ratchet; `llm.py` re-exports the historical names
       ├── usage_accounting.py  ← Append-only physical-model-attempt monetary authority: reserved→dispatched→settled|unresolved (or reserved→released), short cross-process check+append+fsync lock, conservative global/root admission, validated sequence replay/torn-tail quarantine, compatibility projections, and resumable legacy import. Inspectable application candidates also carry exact post-transform raw/context identities and a pre-dispatch manifest/precondition on this same attempt id; specialized boundaries stay honestly opaque
       ├── _usage_rows.py       ← Pure row arithmetic for accounting summaries, limit/integrity decoration, physical-call counts, breakdown buckets, and the exact Skill Review wave/slot projection. It owns no I/O or locks; `usage_accounting.py` re-exports its compatibility names.
       ├── _usage_rows_memo.py  ← In-process read acceleration beside the ledger substrate: the validated-rows memo + fingerprint-keyed render cache for display projections, and the in-lock warm read cache the monetary write paths use (#129). Every cache resumes via the substrate's `LedgerResumeState` fingerprint and falls back to the authoritative full locked read on any doubt; `usage_accounting.py` re-exports its names.
